+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- property `disableScrollToActiveItem` to `PanelSideBarLayout` in order to disable scroll on the active element selection.
+- property `scrollActiveItemToCenter` to `PanelSideBarLayout` in order to scroll the active element to the center of the sidebar.
 
 ## [5.3.0] - 2025-09-30